api/requirements.txt
fastapi
uvicorn[standard]
requests
webdriver-manager
pandas
langchain
langchain-openai
python-dotenv
groq